pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-cure-discovery"
version = "0.1.0"
description = "Accelerating drug discovery through open-source, GPU-powered virtual screening"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Open Cure Discovery Contributors"}
]
keywords = [
    "drug-discovery",
    "molecular-docking",
    "virtual-screening",
    "cheminformatics",
    "machine-learning",
    "gpu-computing",
    "open-science"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    # Core scientific computing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.2.0",

    # Deep learning
    "torch>=2.0.0",
    "onnxruntime-gpu>=1.15.0",

    # Cheminformatics
    "rdkit>=2023.3.1",

    # Database & API
    "sqlalchemy>=2.0.0",
    "requests>=2.28.0",
    "aiohttp>=3.8.0",

    # CLI & UI
    "click>=8.1.0",
    "rich>=13.0.0",
    "typer>=0.9.0",

    # Web interface
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "jinja2>=3.1.0",

    # Configuration
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Utilities
    "tqdm>=4.65.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.4.0",
    "pre-commit>=3.3.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.22.0",
]

visualization = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "py3Dmol>=2.0.0",
]

all = [
    "open-cure-discovery[dev,docs,visualization]",
]

[project.scripts]
ocd = "src.ui.cli.main:app"

[project.urls]
Homepage = "https://github.com/alfredo-baratta/open-cure-discovery"
Documentation = "https://github.com/alfredo-baratta/open-cure-discovery#readme"
Repository = "https://github.com/alfredo-baratta/open-cure-discovery"
Issues = "https://github.com/alfredo-baratta/open-cure-discovery/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
